gcpassignment/requirements.txt
google-cloud-storage
pandas
pymysql
mysql-connector-python
python-dotenv
